Modify EnableSites and DisableSites to use FileManager
Use FileManager to create and remove symbolic links instead of using
shell commands through ShellJS lib.

Also modify what is returned in the response for those endpoints. Only
the modified files are returned in the response. If a site is already
enabled, it will not be returned in the response for the EnableSites
endpoint. Same works for DisableSites.

Signed-off-by: Éverton Arruda <root@arruda.eti.br>